Better placeholder string for Arabic positional forms
Guess placeholder string from glyph name for glyphs without code points.
Checks for the extensions used for Arabic positional forms and add Zero
Width Joiner as appropriate to force the desired positional form.

The glyph-shape-placeholder div needs dir="auto" as ZWJ gets ignored at
string boundaries if text direction is LTR (which is the default in
HTML). With dir="auto", the direction will be resolved based on the
string, so it will be RTL for Arabic text.